Progress: Expired Product Showing and Scheduling bug fix

+++ app/Services/MyBlProductSchedulerService.php
@@ -43,11 +43,11 @@ class MyBlProductSchedulerService
     public function productSchedule()
     {
         $currentTime = Carbon::parse()->format('Y-m-d H:i:s');
-        $products = $this->myblProductRepository->findAll();
+        $products = $this->myblProductRepository->findScheduleProductList();
 
         foreach ($products as $product) {
 
-            $productSchedule = $this->myblProductScheduleRepository->findScheduleDataByProductCode($product['product_code']);
+            $productSchedule = $this->myblProductScheduleRepository->findScheduleDataByProductCodeV2($product['product_code']);
 
             if(is_null($productSchedule)) {
                 continue;
@@ -65,13 +65,15 @@ class MyBlProductSchedulerService
                 if ($product->is_tags_schedule) {
                     $productTags = $this->myblProductTagRepository->findTagIdByProductCode($product['product_code']);
 
-                    if(!is_null($productSchedule->tags)) {
-
-                        //TODO : Need ProductTag Repository to get ProductTag Title
+                    $productScheduleData['tags'] = null;
+                    $productData['tag'] = null;
+                    if(!$productTags->isEmpty()) {
+                        $productScheduleData['tags'] = $productTags;
+                    }
+                    if ($productScheduleData['tags'] != null) {
                         $firstTag = ProductTag::where('id', json_decode($productSchedule->tags)[0])->first();
                         $tag = $firstTag->title;
                         $productData['tag'] = $tag;
-                        $productScheduleData['tags'] = json_encode($productTags);
                         $this->myblProductTagRepository->deleteByProductCode($product['product_code']);
                         $tags = [];
                         foreach (json_decode($productSchedule->tags) as $productScheduleTag) {
@@ -115,6 +117,7 @@ class MyBlProductSchedulerService
                     Log::info($e->getMessage());
                 }
             } elseif ($currentTime > $productSchedule['end_date'] && $productSchedule['change_state_status'] == 1) {
+
                 $productData = [];
                 $productScheduleData = [];
                 if ($product->is_banner_schedule) {
@@ -125,20 +128,27 @@ class MyBlProductSchedulerService
                 if ($product->is_tags_schedule) {
                     //TODO : Need ProductTag Repository to get ProductTag Title
                     $productTags = $this->myblProductTagRepository->findTagIdByProductCode($product['product_code']);
-                    $firstTag = ProductTag::where('id', json_decode($productSchedule->tags)[0])->first();
-                    $tag = $firstTag->title;
-                    $productData['tag'] = $tag;
-                    $productScheduleData['tags'] = $productTags;
-                    $this->myblProductTagRepository->deleteByProductCode($product['product_code']);
-                    $tags = [];
-                    foreach (json_decode($productSchedule->tags) as $productScheduleTag) {
 
-                        $data['product_code'] = $product['product_code'];
-                        $data['product_tag_id'] = $productScheduleTag;
-
-                        $tags [] = $data;
+                    $productScheduleData['tags'] = null;
+                    $productData['tag'] = null;
+                    if(!$productTags->isEmpty()) {
+                        $productScheduleData['tags'] = $productTags;
                     }
-                    $this->myblProductTagRepository->insert($tags);
+                    if ($productScheduleData['tags'] != null) {
+                        $firstTag = ProductTag::where('id', json_decode($productSchedule->tags)[0])->first();
+                        $tag = $firstTag->title;
+                        $productData['tag'] = $tag;
+                        $this->myblProductTagRepository->deleteByProductCode($product['product_code']);
+                        $tags = [];
+                        foreach (json_decode($productSchedule->tags) as $productScheduleTag) {
+
+                            $data['product_code'] = $product['product_code'];
+                            $data['product_tag_id'] = $productScheduleTag;
+
+                            $tags [] = $data;
+                        }
+                        $this->myblProductTagRepository->insert($tags);
+                    }
                 }
 
                 if ($product->is_visible_schedule) {
@@ -156,7 +166,11 @@ class MyBlProductSchedulerService
                     $productScheduleData['base_msisdn_group_id'] = $product['base_msisdn_group_id'];
                 }
                 $productScheduleData['change_state_status'] = 0;
-
+                $productData['is_banner_schedule'] = 0;
+                $productData['is_tags_schedule'] = 0;
+                $productData['is_visible_schedule'] = 0;
+                $productData['is_pin_to_top_schedule'] = 0;
+                $productData['is_base_msisdn_group_id_schedule'] = 0;
                 try {
                     DB::beginTransaction();
 
@@ -180,56 +194,65 @@ class MyBlProductSchedulerService
 
         $productData = [];
         $productScheduleData = [];
-        if ($product->is_banner_schedule) {
+        if ($product->is_banner_schedule && $productSchedule->change_state_status) {
             $productData['media'] = $productSchedule['media'];
             $productScheduleData['media'] = $product['media'];
         }
 
-        if ($product->is_tags_schedule) {
+        if ($product->is_tags_schedule && $productSchedule->change_state_status) {
             //TODO : Need ProductTag Repository to get ProductTag Title
             $productTags = $this->myblProductTagRepository->findTagIdByProductCode($product['product_code']);
-            $firstTag = ProductTag::where('id', json_decode($productSchedule->tags)[0])->first();
-            $tag = $firstTag->title;
-            $productData['tag'] = $tag;
-            $productScheduleData['tags'] = $productTags;
-            $this->myblProductTagRepository->deleteByProductCode($product['product_code']);
-            $tags = [];
-            foreach (json_decode($productSchedule->tags) as $productScheduleTag) {
 
-                $data['product_code'] = $product['product_code'];
-                $data['product_tag_id'] = $productScheduleTag;
-
-                $tags [] = $data;
+            $productScheduleData['tags'] = null;
+            $productData['tag'] = null;
+            if(!$productTags->isEmpty()) {
+                $productScheduleData['tags'] = $productTags;
             }
-            $this->myblProductTagRepository->insert($tags);
+            if ($productScheduleData['tags'] != null) {
+                $firstTag = ProductTag::where('id', json_decode($productSchedule->tags)[0])->first();
+                $tag = $firstTag->title;
+                $productData['tag'] = $tag;
+                $this->myblProductTagRepository->deleteByProductCode($product['product_code']);
+                $tags = [];
+                foreach (json_decode($productSchedule->tags) as $productScheduleTag) {
+
+                    $data['product_code'] = $product['product_code'];
+                    $data['product_tag_id'] = $productScheduleTag;
+
+                    $tags [] = $data;
+                }
+                $this->myblProductTagRepository->insert($tags);
+            }
+
         }
 
-        if ($product->is_visible_schedule) {
+        if ($product->is_visible_schedule && $productSchedule->change_state_status) {
             $productData['is_visible'] = $productSchedule['is_visible'];
             $productScheduleData['is_visible'] = $product['is_visible'];
         }
 
-        if ($product->is_pin_to_top_schedule) {
+        if ($product->is_pin_to_top_schedule && $productSchedule->change_state_status) {
             $productData['pin_to_top'] = $productSchedule['pin_to_top'];
             $productScheduleData['pin_to_top'] = $product['pin_to_top'];
         }
 
-        if ($product->is_base_msisdn_group_id_schedule) {
+        if ($product->is_base_msisdn_group_id_schedule && $productSchedule->change_state_status) {
             $productData['base_msisdn_group_id'] = $productSchedule['base_msisdn_group_id'];
             $productScheduleData['base_msisdn_group_id'] = $product['base_msisdn_group_id'];
         }
+
         $productScheduleData['change_state_status'] = 0;
-        $productScheduleData['start_date'] = null;
-        $productScheduleData['end_date'] = null;
+        $productScheduleData['is_cancel'] = 1;
+        $productData['is_banner_schedule'] = 0;
+        $productData['is_tags_schedule'] = 0;
+        $productData['is_visible_schedule'] = 0;
+        $productData['is_pin_to_top_schedule'] = 0;
+        $productData['is_base_msisdn_group_id_schedule'] = 0;
 
         try {
             DB::beginTransaction();
 
-            if($productSchedule->change_state_status) {
-
-                $this->myblProductRepository->updateDataById($product['id'], $productData);
-            }
-
+            $this->myblProductRepository->updateDataById($product['id'], $productData);
             $this->myblProductScheduleRepository->updateDataById($productSchedule['id'], $productScheduleData);
 
             DB::commit();
